Encapsulate educational modals state and data fetching
