+++ src/components/MoneyReportHeaderEducationalModals.tsx
@@ -1,11 +1,16 @@
 import {shouldFailAllRequestsSelector} from '@selectors/Network';
-import React from 'react';
-import type {OnyxEntry} from 'react-native-onyx';
+import React, {useImperativeHandle, useState} from 'react';
+import type {Ref} from 'react';
 import type {ValueOf} from 'type-fest';
 import useNetwork from '@hooks/useNetwork';
 import useOnyx from '@hooks/useOnyx';
+import usePaginatedReportActions from '@hooks/usePaginatedReportActions';
+import useTransactionsAndViolationsForReport from '@hooks/useTransactionsAndViolationsForReport';
 import {setNameValuePair} from '@libs/actions/User';
+import getNonEmptyStringOnyxID from '@libs/getNonEmptyStringOnyxID';
+import {getAllNonDeletedTransactions} from '@libs/MoneyRequestReportUtils';
 import Navigation from '@libs/Navigation/Navigation';
+import {getFilteredReportActionsForReportView, getOneTransactionThreadReportID, getOriginalMessage, isMoneyRequestAction} from '@libs/ReportActionsUtils';
 import {changeMoneyRequestHoldStatus, rejectMoneyRequestReason} from '@libs/ReportUtils';
 import {dismissRejectUseExplanation} from '@userActions/IOU';
 import CONST from '@src/CONST';
@@ -19,30 +24,52 @@ type RejectModalAction = ValueOf<
     typeof CONST.REPORT.TRANSACTION_SECONDARY_ACTIONS.HOLD | typeof CONST.REPORT.TRANSACTION_SECONDARY_ACTIONS.REJECT | typeof CONST.REPORT.TRANSACTION_SECONDARY_ACTIONS.REJECT_BULK
 >;
 
-type MoneyReportHeaderEducationalModalsProps = {
-    requestParentReportAction: OnyxTypes.ReportAction | null | undefined;
-    transaction: OnyxEntry<OnyxTypes.Transaction>;
-    reportID: string | undefined;
-    isHoldEducationalVisible: boolean;
-    rejectModalAction: RejectModalAction | null;
-    onHoldEducationalDismissed: () => void;
-    onRejectModalDismissed: () => void;
+type MoneyReportHeaderEducationalModalsHandle = {
+    openHoldEducational: () => void;
+    openRejectModal: (action: RejectModalAction) => void;
 };
 
-function MoneyReportHeaderEducationalModals({
-    requestParentReportAction,
-    transaction,
-    reportID,
-    isHoldEducationalVisible,
-    rejectModalAction,
-    onHoldEducationalDismissed,
-    onRejectModalDismissed,
-}: MoneyReportHeaderEducationalModalsProps) {
+type MoneyReportHeaderEducationalModalsProps = {
+    reportID: string | undefined;
+    ref: Ref<MoneyReportHeaderEducationalModalsHandle>;
+};
+
+function MoneyReportHeaderEducationalModals({reportID, ref}: MoneyReportHeaderEducationalModalsProps) {
+    const [isHoldEducationalModalVisible, setIsHoldEducationalModalVisible] = useState(false);
+    const [rejectModalAction, setRejectModalAction] = useState<RejectModalAction | null>(null);
+
     const {isOffline} = useNetwork();
     const [shouldFailAllRequests] = useOnyx(ONYXKEYS.NETWORK, {selector: shouldFailAllRequestsSelector});
 
+    // Fetch report data needed for educational modals
+    const [moneyRequestReport] = useOnyx(`${ONYXKEYS.COLLECTION.REPORT}${reportID}`);
+    const [chatReport] = useOnyx(`${ONYXKEYS.COLLECTION.REPORT}${moneyRequestReport?.chatReportID}`);
+    const {reportActions: unfilteredReportActions} = usePaginatedReportActions(moneyRequestReport?.reportID);
+    const reportActions = getFilteredReportActionsForReportView(unfilteredReportActions);
+    const {transactions: reportTransactions} = useTransactionsAndViolationsForReport(moneyRequestReport?.reportID);
+
+    // Derive transaction thread and parent action
+    const nonDeletedTransactions = getAllNonDeletedTransactions(reportTransactions, reportActions, isOffline, true);
+    const visibleTransactionsForThreadID = nonDeletedTransactions?.filter((t) => isOffline || t.pendingAction !== 'delete');
+    const reportTransactionIDs = visibleTransactionsForThreadID?.map((t) => t.transactionID);
+    const transactionThreadReportID = getOneTransactionThreadReportID(moneyRequestReport, chatReport, reportActions ?? [], isOffline, reportTransactionIDs);
+    const [transactionThreadReport] = useOnyx(`${ONYXKEYS.COLLECTION.REPORT}${transactionThreadReportID}`);
+
+    const requestParentReportAction =
+        reportActions && transactionThreadReport?.parentReportActionID
+            ? reportActions.find((action): action is OnyxTypes.ReportAction<typeof CONST.REPORT.ACTIONS.TYPE.IOU> => action.reportActionID === transactionThreadReport.parentReportActionID)
+            : null;
+
+    const iouTransactionID = isMoneyRequestAction(requestParentReportAction) ? getOriginalMessage(requestParentReportAction)?.IOUTransactionID : undefined;
+    const [transaction] = useOnyx(`${ONYXKEYS.COLLECTION.TRANSACTION}${getNonEmptyStringOnyxID(iouTransactionID)}`);
+
+    useImperativeHandle(ref, () => ({
+        openHoldEducational: () => setIsHoldEducationalModalVisible(true),
+        openRejectModal: (action: RejectModalAction) => setRejectModalAction(action),
+    }));
+
     const dismissModalAndUpdateUseHold = () => {
-        onHoldEducationalDismissed();
+        setIsHoldEducationalModalVisible(false);
         setNameValuePair(ONYXKEYS.NVP_DISMISSED_HOLD_USE_EXPLANATION, true, false, !shouldFailAllRequests);
         if (requestParentReportAction) {
             changeMoneyRequestHoldStatus(requestParentReportAction, transaction, isOffline);
@@ -70,7 +97,7 @@ function MoneyReportHeaderEducationalModals({
                 rejectMoneyRequestReason(requestParentReportAction);
             }
         }
-        onRejectModalDismissed();
+        setRejectModalAction(null);
     };
 
     return (
@@ -81,7 +108,7 @@ function MoneyReportHeaderEducationalModals({
                     onConfirm={dismissRejectModalBasedOnAction}
                 />
             )}
-            {!!isHoldEducationalVisible && (
+            {!!isHoldEducationalModalVisible && (
                 <HoldSubmitterEducationalModal
                     onClose={dismissModalAndUpdateUseHold}
                     onConfirm={dismissModalAndUpdateUseHold}
@@ -91,5 +118,5 @@ function MoneyReportHeaderEducationalModals({
     );
 }
 
-export type {RejectModalAction};
+export type {RejectModalAction, MoneyReportHeaderEducationalModalsHandle};
 export default MoneyReportHeaderEducationalModals;
